step 0
* define super class for SQL translators
* move replicated code to super class
* keep only specific code in crate translator
